Moved UserHelperTest to Altinn.App.Core.Tests, moved AltinnPartyClientMock and ProfileClientMock and related test data to Altinn.App.Tests.Common.
