+++ test/Altinn.App.Tests.Common/Data/Register/Party/500600.json
