+++ test/Altinn.App.Tests.Common/Data/Register/Party/500802.json
