+++ test/Altinn.App.Tests.Common/Mocks/AltinnPartyClientMock.cs
@@ -1,10 +1,10 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using Altinn.App.Api.Tests.Data;
 using Altinn.App.Core.Internal.Registers;
+using Altinn.App.Tests.Common.Data;
 using Altinn.Platform.Register.Models;
 
-namespace Altinn.App.Api.Tests.Mocks;
+namespace Altinn.App.Tests.Common.Mocks;
 
 public class AltinnPartyClientMock : IAltinnPartyClient
 {
@@ -13,7 +13,7 @@ public class AltinnPartyClientMock : IAltinnPartyClient
         Converters = { new JsonStringEnumConverter() },
     };
 
-    private readonly string _partyFolder = TestData.GetAltinnProfilePath();
+    private readonly string _partyFolder = CommonTestData.GetAltinnProfilePath();
 
     public async Task<Party?> GetParty(int partyId)
     {
